implemented some dodgy CI/CD
we build a Docker image on every push to the master or staging branches. this image is then ran on my server, updated constantly via cioban.

this image is tied with the rest of my infrastructure to get webhooks from GitHub for the `master` and `staging` branches.

when the webhook is fired for the appropriate branches, we pull the latest version of the repo, build it with Hugo, and deploy the artifacts to my webserver.